+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Define 'entityTabsAfter' and 'recordTabsAfter' TenantsEditPanel component parameters, thus bumping minor candidate version number
     - TenantsManager is a NodeJS EventEmitter on server side
     - Server-side functions are available in Tenants.s namespace
+    - Open the new tenant dialog on first pane
 
 ### 1.2.0